+++ packages/react-dogfood/pages/guest/join/[guestCallId].tsx
@@ -1,6 +1,6 @@
 import { GetServerSidePropsContext } from 'next';
 import {
-  StreamMeeting,
+  StreamCall,
   StreamVideo,
   useCreateStreamVideoClient,
   UserResponse,
@@ -66,9 +66,9 @@ export default function GuestCallRoom(props: GuestCallRoomProps) {
         <meta name="viewport" content="initial-scale=1.0, width=device-width" />
       </Head>
       <StreamVideo client={client}>
-        <StreamMeeting callId={callId} callType={callType} autoJoin={false}>
+        <StreamCall callId={callId} callType={callType} autoJoin={false}>
           <MeetingUI enablePreview={mode !== 'anon'} />
-        </StreamMeeting>
+        </StreamCall>
       </StreamVideo>
     </>
   );